Update client interface using react library
The current implementation adds react components (movie-card, movie-view, main-view) that displays the list of all movies or details of one movie on a click event. First step to build an SPA that will be connected on the server in the next commits.

Test Plan:
PASS: Open the app in browser: all the movies are displayed
PASS: Click on one movie to display the details
PASS: Click on back button to go to the main menu