Improve the home screen UI
The actual interface displays movie title as simple text, an therefore is not enough appealing.

This commit will use imageUrl to load movie picture in place of the title. To allow this,  a little change will be made to the return function of movie-card component.

The rest remains unchanged, the onClick event is now on the <img > tag added in the movie-card return function.