Connect Client interface to Movie_API
Connects to Client interface:

The client interface displays movie based on harcoded data used as placeholder during client interface building process.

This commit remove those data and connects the client application to heroku api and allow client app to fetch data on Atlas DB using the API.

The main change is the addition of the useEffect hook to fetch movie from API.
The authentication to API is temporally disabled on movie endpoint on the server side (movie_api) for testing purposes

Enforce Props transmission rules via components.
It is not mandatory to tranmit all the props required to lower components.

This commit enforces rules on object and fonctions to be passed between components.
PropTypes library is imported in MovieCard and MovieView components, and static properties are set on each passed object and function.